Initializing Tic-Tac-Toe
This is a personal project to give me experience with C++, game development,
and some things to finally start building a portfolio as a programmer. This
program will be released as Open Source under the 2 clause BSD License.

This project will probably keep me occupied at least through Spring Break. My
goal is to get the terminal version of the game completed by the end of March.
3D should be an extendable interface for 2D if I understand C++ correctly. If
not, then there may be a delay in the creation of a 3D board.