Added the isTerminal function.
This function will see if the board is at a terminal state. The first check is
along the diagonal form the top left corner to the bottom right corner. The
next check is along the diagonal form the bottom left corner to the top right
corner. The next check will see if any row is completed by a single player. The
next check will see if any column is completed by a single player. The last
check will see if the board is completely full.

This function will not return who the winner of the game is. It will merely
check if the game is over or not.